Exclude test tasks from the production environment. Use a setting for creating the password with baseline data

+++ Rakefile
@@ -1,45 +1,53 @@
 require 'rake'
-require 'rspec/core/rake_task'
+require File.join(File.dirname(__FILE__), 'app', 'hansel')
 
-desc "Run all specs"
-task :spec => ["spec:controllers", "spec:models"]
-task :test => [:spec]
-task :default => [:spec]
+if !ENV['RACK_ENV'] || ["test", "development"].include?(ENV['RACK_ENV'])
 
-namespace "spec" do
+	require 'rspec/core/rake_task'
 
-	RSpec::Core::RakeTask.new(:controllers) do |t|
-		t.pattern = "./spec/controllers/*_spec.rb"
+	desc "Run all specs"
+	task :spec => ["spec:controllers", "spec:models"]
+	task :test => [:spec]
+	task :default => [:spec]
+
+	namespace "spec" do
+
+		RSpec::Core::RakeTask.new(:controllers) do |t|
+			t.pattern = "./spec/controllers/*_spec.rb"
+		end
+
+		RSpec::Core::RakeTask.new(:models) do |t|
+			t.pattern = "./spec/models/*_spec.rb"
+		end
+
 	end
 
-	RSpec::Core::RakeTask.new(:models) do |t|
-		t.pattern = "./spec/models/*_spec.rb"
+
+	require File.join(File.dirname(__FILE__), 'spec', 'factories', 'gist_factory')
+
+	namespace "db" do
+
+		desc "Seed the database"
+		task :seed do
+			60.times { FactoryGirl.create(:gist) }
+			43.times { FactoryGirl.create(:text_gist) }
+			32.times { FactoryGirl.create(:done_gist) }
+		end
+
+		desc "Remove all test data"
+		task :drop do
+			Mongoid.master.collections.select { |c| c.name !~ /system/ }.each(&:drop)
+		end
+
+		desc "Drop, reseed and insert baseline data"
+		task :rebuild => ["db:drop", "db:seed", "db:baseline"]
 	end
 
 end
 
-require File.join(File.dirname(__FILE__), 'app', 'hansel')
-require File.join(File.dirname(__FILE__), 'spec', 'factories', 'gist_factory')
-
 namespace "db" do
-
-	desc "Seed the database"
-	task :seed do
-		60.times { FactoryGirl.create(:gist) }
-		43.times { FactoryGirl.create(:text_gist) }
-		32.times { FactoryGirl.create(:done_gist) }
-	end
-
-	desc "Remove all test data"
-	task :drop do
-		Mongoid.master.collections.select { |c| c.name !~ /system/ }.each(&:drop)
-	end
-
 	desc "Insert baseline data"
 	task :baseline do
 		require File.join(File.dirname(__FILE__), 'app', 'baseline_data')
 	end
-
-	desc "Drop, reseed and insert baseline data"
-	task :rebuild => ["db:drop", "db:seed", "db:baseline"]
 end
